20150202: Update rlw-stub.sh
Bring a bunch of changes from rlw.sh to rlw-stub.sh, now includes GNU GPL info as well.

+++ rlw-stub.sh
@@ -1,69 +1,40 @@
 #!/usr/bin/env bash
-download () {
-	wget $1 -O $2 2>&1 | \
-	sed -u 's/.* \([[0-9]]\+%\)\ \+\([[0-9.]]\+.\) \(.*\)/\1\n# Downloading at \2\/s, ETA \3/' | \
-	zenity \
-		--progress \
-		--window-icon=$RBXICON \
-		--title='Downloading' \
-		--auto-close \
-		--no-cancel \
-		--width=362 \
-		--height=122
+#
+#  Copyright 2015 Jonathan Alfonso <alfonsojon1997@gmail.com>
+#
+#  This program is free software; you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation; either version 2 of the License, or
+#  (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program; if not, write to the Free Software
+#  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
+#  MA 02110-1301, USA.
+#
+#
+
+rwget () {
+	wget "$@" 2>&1 | sed -u 's/.* \([0-9]\+%\)\ \+\([0-9.]\+.\) \(.*\)/\1\n# Downloading at \2\/s, ETA \3/' | zenity --progress --window-icon="$RBXICON" --title='Downloading' --auto-close --no-cancel --width=450 --height=120; [[ $? = "0" ]] || { spawndialog error "wget download failed. \nSee terminal for details. (exit code $?)"; exit $?; }
 }
 spawndialog () {
-	zenity \
-		--window-icon=$RBXICON \
-		--title='Roblox Linux Wrapper' \
-		--$1 \
-		--no-wrap \
-		--text="$2" \
-		--width=362 \
-		--height=122
+	echo -e "$2"
+	zenity --window-icon="$RBXICON" --title='Roblox Linux Wrapper v'"$RLWVERSION"'-'"$RLWCHANNEL" --"$1" --no-wrap --text="$2"
 }
-depcheck () {
-	if command -v zenity >/dev/null 2>&1; then
-		echo 'zenity installed, continuing'
-	else
-		echo "Please install zenity via your system's package manager."
-		exit 127
-	fi
-	if command -v shasum >/dev/null 2>&1; then
-		echo 'shasum installed, continuing'
-	else
-		echo "Please install shasum via your system's package manager."
-		spawndialog error "Please install shasum via your system's package manager"
-		exit 127
-	fi
-	if command -v wget >/dev/null 2>&1; then
-		echo 'wget installed, continuing'
-	else
-		echo "Please install wget via your system's package manager."
-		spawndialog error "Please install wget via your system's package manager."
-		exit 127
-	fi
-}
-
-depcheck
-
-if [[ -e $HOME/.local/share/icons/hicolor/512x512/apps/roblox.png ]]; then
-	export RBXICON=$HOME/.local/share/icons/hicolor/512x512/apps/roblox.png
-else
-	download http://img1.wikia.nocookie.net/__cb20130302012343/robloxhelp/images/f/fb/ROBLOX_Circle_Logo.png $HOME/.local/share/icons/hicolor/512x512/apps/roblox.png
-	export RBXICON=$HOME/.local/share/icons/hicolor/512x512/apps/roblox.png
-fi
-
+[[ -e $(which zenity) && $(which shasum) && $(which wget)  ]] || { spawndialog error "Missing dependencies! Make sure zenity, wget, wine, and wine-staging are installed."; exit 1; }
 if [[ ! -e $HOME/.rlw/rlw.sh ]]; then
 	download https://raw.githubusercontent.com/alfonsojon/roblox-linux-wrapper/master/rlw.sh $HOME/.rlw/rlw.sh
 	cp $HOME/.rlw/rlw.sh $HOME/.rlw/rlw.sh.update
 fi
-
-download https://raw.githubusercontent.com/alfonsojon/roblox-linux-wrapper/master/rlw.sh $HOME/.rlw/rlw.sh.update
-
+rwget https://raw.githubusercontent.com/alfonsojon/roblox-linux-wrapper/master/rlw.sh -O $HOME/.rlw/rlw.sh.update
 if [[ `shasum $HOME/.rlw/rlw.sh.update | cut -d' ' -f1` != `cat $HOME/.rlw/update.ignored | cut -d' ' -f1` ]]; then
 	rm -rf $HOME/.rlw/update.ignored
 fi
-
 if [[ `shasum $HOME/.rlw/rlw.sh | cut -d' ' -f1` != `shasum $HOME/.rlw/rlw.sh.update | cut -d' ' -f1` ]]; then
 	if [[ ! -e $HOME/.rlw/update.ignored ]]; then
 		if [[ `cat $HOME/.rlw/update.ignored | cut -d' ' -f1` != `shasum $HOME/.rlw/rlw.sh.update` ]]; then
@@ -78,7 +49,6 @@ if [[ `shasum $HOME/.rlw/rlw.sh | cut -d' ' -f1` != `shasum $HOME/.rlw/rlw.sh.up
 	fi
 fi
 
-echo 'Loading rlw.sh ...'
-echo ''
+echo -e "Loading rlw.sh ... \n"
 chmod +x $HOME/.rlw/rlw.sh
 bash $HOME/.rlw/rlw.sh
